refactor: User쪽 네이밍 변경 및 코드정리 마무리
