+++ src/main/java/bit/user/dto/UserCreateRequest.java
@@ -1,7 +1,7 @@
 package bit.user.dto;
 
 import bit.user.domain.User;
-import bit.user.oauth.OauthPlatformStatus;
+import bit.user.oauth.enums.OauthPlatformStatus;
 import bit.user.oauth.kakao.domain.KakaoUserInfo;
 import java.time.LocalDateTime;
 import lombok.Builder;
