--- src/main/java/bit/user/oauth/OauthPlatformStatus.java
@@ -1,9 +1,0 @@
-package bit.user.oauth;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-public enum OauthPlatformStatus {
-    KAKAO, NAVER, GOOGLE;
-
-}
