+++ src/main/java/bit/user/oauth/enums/OauthPlatformStatus.java
@@ -1,0 +1,5 @@
+package bit.user.oauth.enums;
+
+public enum OauthPlatformStatus {
+    KAKAO, NAVER, GOOGLE;
+}
