+++ src/test/java/bit/user/service/UserServiceImplTest.java
@@ -1,6 +1,6 @@
 package bit.user.service;
 
-import static bit.user.oauth.OauthPlatformStatus.KAKAO;
+import static bit.user.oauth.enums.OauthPlatformStatus.KAKAO;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
